+++ dev/serve.vue
@@ -247,6 +247,10 @@
         />
         <hr />
 
+        <h2>SectionSpacesList Component</h2>
+        <section-spaces-list :items="mockSectionSpacesList.items" />
+        <hr />
+
         <h2>SmartLink Component</h2>
         <smart-link to="/"> This will render as a vue-router link </smart-link>
         <hr />
@@ -254,6 +258,16 @@
 
         <h2>FlexibleBlock Components</h2>
         <flexible-blocks :blocks="flexibleBlocks" />
+
+        <hr />
+        <hr />
+
+        <h2>Simple Cards</h2>
+        <simple-cards
+            :items="mockSimpleCards.cards"
+            :section-title="mockSimpleCards.sectionTitle"
+            :section-summary="mockSimpleCards.sectionSummary"
+        />
     </div>
 </template>
 
@@ -516,6 +530,31 @@ export default Vue.extend({
                         prompt: "Velit Exceptur",
                         dates: "Datum 1 - Datum 18, 2021",
                         to: "/visit/uri/2",
+                    },
+                ],
+            },
+            mockSectionSpacesList: {
+                items: [
+                    {
+                        to: "https://calendar.library.ucla.edu",
+                        title: "Internal Portico",
+                        location: "Fames Turpis Building",
+                        text: "Etiam faucibus tellus quis magna finibus auctor. Fusce sagittis vel lectus sed ullamcorper. Proin faucibus nibh ut purus consequat lobortis. Aenean eu fermentum leo, vitae egestas eros. Vivamus urna nibh, porttitor in egestas eu, eleifend ac felis. Praesent sed tristique nibh. Nulla dignissim purus at metus consequat, vitae pharetra orci efficitur. Vivamus facilisis massa vel ipsum vestibulum, porta efficitur dolor aliquam.",
+                        buttonUrl: "http://google.com",
+                    },
+                    {
+                        to: "https://www.gettyimages.com/photos/cute-llamas",
+                        title: "External Von Junz",
+                        location: "Debordist Room",
+                        text: "If one examines Marxist class, one is faced with a choice: either reject Baudrillardist simulation or conclude that the significance of the artist is social comment. The premise of Debordist situation implies that the law is capable of significance. Therefore, any number of constructions concerning a self-sufficient reality may be found.",
+                        buttonUrl: "http://google.com",
+                    },
+                    {
+                        to: "",
+                        title: "Null Beefeater Numpty",
+                        location: "Pennyboy Cottage",
+                        text: "Gosh beefeater numpty scouser rambunctious a bottle of plonk know your onions, cornish pasty spend a penny cheerio down the local mush. What a mug one off blimey a comely wench nose rag ponce gobsmacked pennyboy cottage pie, ey up fake tan spam fritters marvelous chippy hedgehog chav.",
+                        buttonUrl: "http://google.com",
                     },
                 ],
             },
@@ -878,6 +917,33 @@ export default Vue.extend({
         },
         parsedNavSecondary() {
             return [...API.links, { ...API.links[0] }]
+        },
+        mockSimpleCards() {
+            return {
+                id: "13783",
+                sectionTitle: "Fix Potholes",
+                sectionSummary:
+                    "<p>Get pothole repair service in Los Angeles and surrounding areas. We look forward to serving all your needs. Contact us now for fast service. Get A Free Quote.</p>",
+                cards: [
+                    {
+                        id: "6622",
+                        text: "<p>Learn how to use a VPN or proxy server to access Library resources with your UCLA Logon ID through a series of helpful videos.</p>",
+                        title: "Connecting from Off-Campus",
+                        slug: "connecting-from-off-campus",
+                        uri: "help/services-resources/connecting-from-off-campus",
+                        type: "resource",
+                    },
+                    {
+                        id: "6407",
+                        text: "<p>A research guide</p>",
+                        title: "A new research guide",
+                        slug: "a-new-research-guide",
+                        uri: null,
+                        externalResourceUrl:
+                            "http://libguides.library.ucla.edu",
+                    },
+                ],
+            }
         },
     },
 })